Rename `near` and `far`.
This is necessary because of global namespace pollution in `windef.h`
for people that wish to use this library the Win32 API at the same time.
Workarounds exist, like including files in a different order, or
undefining `near` and `far` before including this library, but it is a
common enough problem that it makes sense to change the library, as this
is such a minor change.

Should fix #64 and #62.