Fix issue #89
This simply makes all `Index` setters take `size_t` instead of `Index::Type`. The added benefit is
that the cast is done once inside the `Index` class, instead of being done everywhere the setter is
called. Thus, the check that the primitive count and first index value are representable also
happens once in the same place.